Add MIME type for `.torrent` files
Serve `.torrent` files with the `application/x-bittorrent` MIME type
instead of the default `application/octet-stream`. This helps improve
the user experience in some browsers (e.g.: in Firefox, the user is
offered the option to open the file with a torrent application, if
it has one installed).

Close: h5bp/server-configs-nginx#4.