Add configs for cursor images (`.cur`)
Apply all the configs used for `.ico` files to `.cur` files as the two
image formats are almost identical.

The MIME type for the `.cur` file doesn't matter as browsers seem to
sniff the content. However, `image/x-icon` is specified because it's
the one usually used for `.ico` file and because it's shorther then the
other frequentlly used one (`image/vnd.microsoft.icon`), making the
header size smaller.

See also:
* http://en.wikipedia.org/wiki/ICO_%28file_format%29
* http://msdn.microsoft.com/en-us/library/windows/desktop/ms648007%28v=vs.85%29.aspx

Close: #5.