Updating the configuration for cross-domain with a small example and a resource hook
The biggest issue that's here is that * is being used which will allow any domain to use XHR connections to your domain. This opens up many security issues as it breaks the same-domain model built into the browser. It's suggested that users instead whitelist domains they wish to give connections to.

Adding further details and links regarding nginx configurations

Organizing the notes; Making sure it can still be used a solid, primer snippet

Took at look at https://github.com/h5bp/server-configs-apache/blob/master/.htaccess again and added the same allowed resources as the default allow-origin.
Not having a wide open allow-origin is the point to the changes.
The notes for whitelisting a domain are there for more advanced usages in domains.

Killing the indenting and blank lines

Quality reading experiences.