Organizing the notes; Making sure it can still be used a solid, primer snippet
Took at look at https://github.com/h5bp/server-configs-apache/blob/master/.htaccess again and added the same allowed resources as the default allow-origin. 
Not having a wide open allow-origin is the point to the changes.
The notes for whitelisting a domain are there for more advanced usages in domains.