Use specific MIME type for '.opus' files
Files with the `.opus` extension should be served with the MIME type `audio/ogg; codecs=opus` due to the fact that they are not simply Ogg Vorbis files, but will likely be treated that way in older browsers when they are served with the content-type they currently have. The single-quotes surrounding the content-type are necessary to escape the semicolon in it; I tried using a backslash, but it doesn't work.